🎨 Move type traits into their own module (+ homogenize trait bounds).
